+++ lib/src/fragment.dart
@@ -4,8 +4,8 @@ import 'package:fragment/src/utils.dart';
 /// A widget to cache [builder] with [deps]
 /// The builder will be called only when [deps] is different (not shallowly equal)
 /// from the previous [deps].
-class Fragment<T extends Widget> extends StatelessWidget {
-  final T Function(BuildContext context) builder;
+class Fragment<T extends Widget> extends StatefulWidget {
+  final T Function(BuildContext context, T prev) builder;
   final Iterable deps;
 
   const Fragment(this.builder, {this.deps = const [], Key key})
@@ -23,7 +23,15 @@ class Fragment<T extends Widget> extends StatelessWidget {
   int get hashCode => deps.hashCode;
 
   @override
-  T build(BuildContext context) {
-    return builder(context);
+  _FragmentState createState() => _FragmentState();
+}
+
+class _FragmentState<T extends Widget> extends State<Fragment<T>> {
+  T prev;
+
+  @override
+  Widget build(BuildContext context) {
+    prev = widget.builder(context, prev);
+    return prev;
   }
 }